backend/requirements.txt
fastapi
uvicorn
sqlalchemy
aiofiles
python-multipart
redis
websockets
psycopg2-binary
python-dotenv
